+++ tests/DataStorage/Cache/Connection/RedisCacheTest.php
@@ -45,6 +45,10 @@ final class RedisCacheTest extends \PHPUnit\Framework\TestCase
 
     protected function tearDown() : void
     {
+        if (!isset($this->cache)) {
+            return;
+        }
+
         $this->cache->flushAll();
     }
